129 - move graph.py out of puml_graph folder
